Update navigationMecanumPID.java
Now rotateToAngle works and loopPID takes a 3rd argument, TpTurn. This does not affect older programs, but simply uses the Tp in case 3 and 4. Older programs will default to 0.5 or -0.5 depending on direction. cases 1, 2, 5, and 6 auto chooses TpTurn from TpSlide and TpForwards, whichever the case uses. moveNoStop chooses whichever is larger, and rotateToAngle defaults to 0.5.